Autoicrement OFF

+++ EPaczucha.core/DTO/PackagePriceDto.cs
@@ -4,7 +4,7 @@
     {
         public int Id { get; set; }
         public decimal Net { get; set; }
-        public int VAT { get; set; }
+        public decimal VAT { get; set; }
         public decimal Gross { get; set; }
     }
 }
