+++ EPaczucha.database/Entities/PackagePrice.cs
@@ -3,7 +3,7 @@
     public class PackagePrice : BaseEntity
     {
         public decimal Net { get; set; }
-        public int VAT { get; set; }
+        public decimal VAT { get; set; }
         public decimal Gross { get; set; }
     }
 }
